+++ Unnamed/src/ObjectPool.hpp
@@ -9,7 +9,7 @@ template <class TObject>
 class ObjectPool
 {
 public:
-	static const unsigned int POOL_SIZE = 100; // Upper limit - 15000@~60fps
+	static const size_t POOL_SIZE = 100; // Upper limit - 15000@~60fps
 
 private:
 	TObject* _firstAvailable;
@@ -18,7 +18,7 @@ private:
 public:
 	ObjectPool();
 	~ObjectPool();
-	TObject& GetObject(const unsigned int& position);
+	TObject& GetObject(const size_t& position);
 	void Create(sf::Texture& texture, WayPoint* wps, const sf::Vector2f spawnPos);
 	void Update(const float& deltaTime);
 	void Render(const std::unique_ptr<sf::RenderWindow>& rw, const float& deltaTime, const float& interpolation);
@@ -30,7 +30,7 @@ ObjectPool<TObject>::ObjectPool()
 	// Free-list implementation
 	_firstAvailable = &_entity[0];
 
-	for (int i = 0; i < POOL_SIZE - 1; i++)
+	for (size_t i = 0; i < POOL_SIZE - 1; i++)
 	{
 		_entity[i].SetNext(&_entity[i + 1]);
 	}
@@ -43,7 +43,7 @@ ObjectPool<TObject>::~ObjectPool()
 {}
 
 template<class TObject>
-TObject& ObjectPool<TObject>::GetObject(const unsigned int& position)
+TObject& ObjectPool<TObject>::GetObject(const size_t& position)
 {
 	return _entity[position];
 }
@@ -64,7 +64,7 @@ void ObjectPool<TObject>::Update(const float& deltaTime)
 	// If entity is in use, update animation and physics
 	// Once entity finished its lifetime return true and set unused entity as head and append to free-list
 	// If entity is not in use skip update
-	for (unsigned int i = 0; i < POOL_SIZE; i++)
+	for (size_t i = 0; i < POOL_SIZE; i++)
 	{
 		if (_entity[i].Update(deltaTime))
 		{
@@ -77,7 +77,7 @@ void ObjectPool<TObject>::Update(const float& deltaTime)
 template<class TObject>
 void ObjectPool<TObject>::Render(const std::unique_ptr<sf::RenderWindow>& rw, const float& deltaTime, const float& interpolation)
 {
-	for (unsigned int i = 0; i < POOL_SIZE; i++)
+	for (size_t i = 0; i < POOL_SIZE; i++)
 	{
 		_entity[i].Render(rw, deltaTime, interpolation);
 	}
